requirements.txt
Flask==2.3.3
psycopg==3.1.18
psycopg-pool==3.2.1
python-dotenv==1.0.0